+++ lib/modules/home/settings_page.dart
@@ -192,6 +192,14 @@ class _SettingsPageState extends State<SettingsPage> {
                   onTap: () =>
                       launch('https://testflight.apple.com/join/HSyZttrr'),
                 ),
+              ListTile(
+                title: Text(LocalizedText(
+                        chs: '更新历史', jpn: '更新履歴', eng: 'Update History')
+                    .localized),
+                onTap: () {
+                  launch('$kProjectHomepage/blob/master/CHANGELOG.md');
+                },
+              ),
               if (Platform.isIOS || Platform.isMacOS)
                 ListTile(
                   title: Text(S.of(context).about_appstore_rating),