+++ lib/modules/shared/quest_card.dart
@@ -67,7 +67,8 @@ class _QuestCardState extends State<QuestCard> {
     String? lastPlace, lastPlaceJp;
     for (int i = 0; i < battles.length; i++) {
       final battle = battles[i];
-      String? place = battle.place?.isNotEmpty==true ? battle.place : lastPlace;
+      String? place =
+          battle.place?.isNotEmpty == true ? battle.place : lastPlace;
       String? placeJp =
           battle.placeJp?.isNotEmpty == true ? battle.placeJp : lastPlaceJp;
       lastPlace = place;
@@ -156,6 +157,23 @@ class _QuestCardState extends State<QuestCard> {
     return children;
   }
 
+  final _clasIcons = {
+    '剑': 'Saber',
+    '弓': 'Archer',
+    '枪': 'Lancer',
+    '骑': 'Rider',
+    '术': 'Caster',
+    '杀': 'Assassin',
+    '狂': 'Berserker',
+    '仇': 'Avenger',
+    // '月外分盾'
+  };
+
+  Widget _getClassIcon(String clsName) {
+    if (!_clasIcons.containsKey(clsName)) return Container();
+    return db.getIconImage('金卡${_clasIcons[clsName]}', width: 16);
+  }
+
   Widget _buildWave(List<Enemy?> enemies) {
     List<Widget> enemyWidgets = enemies.map((enemy) {
       if (enemy == null) return Container();
@@ -167,12 +185,24 @@ class _QuestCardState extends State<QuestCard> {
         if (name?.isNotEmpty == true)
           lines.add(AutoSizeText(name!,
               maxFontSize: 14, maxLines: 1, textAlign: TextAlign.center));
-        lines.add(AutoSizeText('${enemy.className[i]} ${enemy.hp[i]}',
-            maxFontSize: 12,
-            // ensure HP is shown completely
-            minFontSize: 4,
-            maxLines: 1,
-            textAlign: TextAlign.center));
+        lines.add(Row(
+          mainAxisSize: MainAxisSize.min,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            if (_clasIcons.containsKey(enemy.className[i]))
+              _getClassIcon(enemy.className[i]),
+            Flexible(
+              child: AutoSizeText(
+                '${enemy.className[i]} ${enemy.hp[i]}',
+                maxFontSize: 12,
+                // ensure HP is shown completely
+                minFontSize: 1,
+                maxLines: 1,
+                textAlign: TextAlign.center,
+              ),
+            )
+          ],
+        ));
       }
       return Column(
         mainAxisSize: MainAxisSize.min,